fix: restructure types
